finances/requirements.txt
Django
https://github.com/jessemiller/HamlPy/tarball/master
scss
django-registration